Change logic to remove decimal part from Counter (#496)
Math.floor(n) returns integer n which is smaller or equal to n. But in
the case of Long library, which is used in JS SDK for purpose of
handling large numbers, it only removes decimal points. It works the
same way in Go SDK.

Therefore, this commit replaces Math.floor with remove the decimal.

---------

Co-authored-by: Youngteac Hong <susukang98@gmail.com>